# Add an ability to skip tests matching a regexp