--- a/README.md
+++ b/README.md
@@ -9,7 +9,9 @@ by adding
 - support to (de-)select or skip parametrized tests without needing to specify test instance qualifiers
 - support for blank and comment lines in the selection files
 - better integration with the `pytest-xdist`, plugin warning and error messages are passed to the master node with proper stdout or stderr outputs
-- an ability to skip tests matching a certain regexp (the skipline should start with `@regexp:`)
+- an ability to select/skip parameters combinations matching a certain regular expression.
+Put your regexp in the square brackets as a raw string and end the line with `@regexp` suffix:
+`file.py::test[r"REGEXP"]@regexp`
 
 
 Usage
@@ -41,7 +43,8 @@ Example::
     test_parametrized[1]
     test_parametrized
     tests/test_foo.py::test_other
-    @regexp:test_parametrized_complex\[[8|16]-.*-.*\]
+    test_parametrized_complex[r"[8|16]-.*-.*"]@regexp
+    tests/test_foo.py::test_params[r"int32-.*-.*"]@regexp
 
     $~ pytest --select-from-file selection.txt
     $~ pytest --deselect-from-file selection.txt