@anmyachev 2025-07-29
@vlad-penkin I think we should try it.

> If there are a lot of defined regexp'es it could in theory slowdown the skipping process.

If it suddenly starts to take up a significant amount of time, we can always return to the previous method via `--collect-only`.